refactor: simplify of dangling validators
